[1/2] refactor code and project structure
