+++ server/plugin.go
@@ -17,6 +17,8 @@ const (
 )
 
 var (
+	badgeFailedURL     string
+	badgePassedURL     string
 	buildFailedIconURL string
 	buildGreenIconURL  string
 )
