+++ server/store/subscriptions.go
@@ -1,22 +1,44 @@
 package store
 
 import (
+	"fmt"
 	"sort"
 
-	"github.com/nathanaelhoun/mattermost-plugin-circleci/server/circle"
+	"github.com/mattermost/mattermost-server/v5/model"
+	"github.com/nathanaelhoun/mattermost-plugin-circleci/server/utils"
 )
 
 const (
 	subscriptionsKVKey = "subscriptions"
 )
 
+type Subscription struct {
+	ChannelID  string            `json:"ChannelID"`
+	CreatorID  string            `json:"CreatorID"`
+	Flags      SubscriptionFlags `json:"Flags"`
+	Owner      string            `json:"Owner"`
+	Repository string            `json:"Repository"`
+}
+
 type Subscriptions struct {
 	Repositories map[string][]*Subscription
 }
 
-// Add a new subscription in the struct
+func (s *Subscription) ToSlackAttachmentField(username string) *model.SlackAttachmentField {
+	return &model.SlackAttachmentField{
+		Title: utils.GetFullNameFromOwnerAndRepo(s.Owner, s.Repository),
+		Short: true,
+		Value: fmt.Sprintf(
+			"Subscribed by: @%s\nFlags: ` %s`",
+			username,
+			s.Flags.String(),
+		),
+	}
+}
+
+// AddSubscription adds a new subscription in the struct
 func (s *Subscriptions) AddSubscription(newSub *Subscription) {
-	key := circle.GetFullNameFromOwnerAndRepo(newSub.Owner, newSub.Repository)
+	key := utils.GetFullNameFromOwnerAndRepo(newSub.Owner, newSub.Repository)
 
 	repoSubs := s.Repositories[key]
 
@@ -39,10 +61,10 @@ func (s *Subscriptions) AddSubscription(newSub *Subscription) {
 	}
 }
 
-// Remove a subscription from the struct
+// RemoveSubscription removes a subscription from the struct
 // Return true if the subscription has been found and removed
 func (s *Subscriptions) RemoveSubscription(channelID, owner, repository string) bool {
-	key := circle.GetFullNameFromOwnerAndRepo(owner, repository)
+	key := utils.GetFullNameFromOwnerAndRepo(owner, repository)
 
 	repoSubs := s.Repositories[key]
 	if repoSubs == nil {
@@ -91,7 +113,7 @@ func (s *Subscriptions) GetSubscriptionsByChannel(channelID string) []*Subscript
 
 // Return a list of subscribed channel IDs
 func (s *Subscriptions) GetSubscribedChannelsForRepository(owner, repository string) []string {
-	key := circle.GetFullNameFromOwnerAndRepo(owner, repository)
+	key := utils.GetFullNameFromOwnerAndRepo(owner, repository)
 
 	subs := s.Repositories[key]
 	if subs == nil {
